refactor: rename packages from test names to prod. ready names
