+++ src/main/java/com/epam/triangful/TriangFulApplication.java
@@ -1,4 +1,4 @@
-package com.labs.restservice;
+package com.epam.triangful;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,10 +7,10 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
 @EnableSwagger2
-public class RestServiceApplication {
+public class TriangFulApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(RestServiceApplication.class, args);
+        SpringApplication.run(TriangFulApplication.class, args);
     }
 
 }
