+++ src/main/java/com/epam/triangful/calculations/CalculationService.java
@@ -1,7 +1,7 @@
-package com.labs.restservice.calculations;
+package com.epam.triangful.calculations;
 
-import com.labs.restservice.exception.InternalException.InternalArithmeticException;
-import com.labs.restservice.triangle.Triangle;
+import com.epam.triangful.exception.InternalException.InternalArithmeticException;
+import com.epam.triangful.triangle.Triangle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
