+++ src/main/java/com/epam/triangful/triangle/TriangleController.java
@@ -1,20 +1,16 @@
-package com.labs.restservice.triangle;
+package com.epam.triangful.triangle;
 
-import com.labs.restservice.cache.TriangleCacheService;
-import com.labs.restservice.calculations.CalculationResults;
-import com.labs.restservice.calculations.CalculationService;
-import com.labs.restservice.concurrency.ServiceAccessManager;
-import com.labs.restservice.exception.ApiException.ApiRequestException;
+import com.epam.triangful.calculations.CalculationResults;
+import com.epam.triangful.calculations.CalculationService;
+import com.epam.triangful.concurrency.ServiceAccessManager;
+import com.epam.triangful.exception.ApiException.ApiRequestException;
+import com.epam.triangful.cache.TriangleCacheService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Map;
-import java.util.Optional;
 
 @RestController
 public class TriangleController {
